refactor: 피드백 반영 - message 상수 파일 분리 - path 상수 파일 분리
