+++ frontend/src/@constants/api.ts
@@ -28,3 +28,17 @@ export const SEARCH_PARAMS = {
   SEARCH_KEYWORD: "keyword",
   SEARCH_CHANNEL_IDS: "channelIds",
 } as const;
+
+export const ERROR_CODE = {
+  MEMBER_NOT_FOUND: "MEMBER_NOT_FOUND",
+  INVALID_TOKEN: "INVALID_TOKEN",
+  CHANNEL_NOT_FOUND: "CHANNEL_NOT_FOUND",
+  SUBSCRIPTION_DUPLICATE: "SUBSCRIPTION_DUPLICATE",
+  SUBSCRIPTION_INVALID_ORDER: "SUBSCRIPTION_INVALID_ORDER",
+  SUBSCRIPTION_NOT_EXIST: "SUBSCRIPTION_NOT_EXIST",
+  SUBSCRIPTION_ORDER_DUPLICATE: "SUBSCRIPTION_ORDER_DUPLICATE",
+  BOOKMARK_DELETE_FAILURE: "BOOKMARK_DELETE_FAILURE",
+  SUBSCRIPTION_NOT_FOUND: "SUBSCRIPTION_NOT_FOUND",
+  BOOKMARK_NOT_FOUND: "BOOKMARK_NOT_FOUND",
+  MESSAGE_NOT_FOUND: "MESSAGE_NOT_FOUND",
+} as const;
