+++ frontend/src/@constants/path.ts
@@ -1,0 +1,9 @@
+export const PATH_NAME = {
+  HOME: "/",
+  FEED: "/feed",
+  BOOKMARK: "/bookmark",
+  REMINDER: "/reminder",
+  ADD_CHANNEL: "/add-channel",
+  CERTIFICATION: "/certification",
+  SEARCH_RESULT: "/search-result",
+} as const;
