+++ frontend/src/pages/AddChannel/index.tsx
@@ -5,7 +5,7 @@ import Button from "@src/components/@shared/Button";
 import useGetChannels from "@src/hooks/@query/useGetChannels";
 import useMutateChannels from "@src/hooks/@query/useMutateChannels";
 
-import { PATH_NAME } from "@src/@constants";
+import { PATH_NAME } from "@src/@constants/path";
 import { FlexColumn } from "@src/@styles/shared";
 
 import * as Styled from "./style";
