+++ frontend/src/pages/NotFound/index.tsx
@@ -2,7 +2,7 @@ import { Link } from "react-router-dom";
 
 import Button from "@src/components/@shared/Button";
 
-import { PATH_NAME } from "@src/@constants";
+import { PATH_NAME } from "@src/@constants/path";
 import { FlexColumn } from "@src/@styles/shared";
 
 import * as Styled from "./style";
